Split exchange collection from service
