+++ src/internal/m365/collection/exchange/attachment.go
