+++ src/internal/m365/collection/exchange/attendees.go
