+++ src/internal/m365/collection/exchange/collection.go
