+++ src/internal/m365/collection/exchange/consts.go
