+++ src/internal/m365/collection/exchange/container_resolver.go
