+++ src/internal/m365/collection/exchange/mail_container_cache.go
