+++ src/internal/m365/collection/exchange/transform.go
